func setValue of bar only accept an int parameter
fix: use int parameter for bar.setValue in scrollRequest

+++ PPOCRLabel/PPOCRLabel.py
@@ -1422,7 +1422,7 @@ class MainWindow(QMainWindow):
     def scrollRequest(self, delta, orientation):
         units = - delta / (8 * 15)
         bar = self.scrollBars[orientation]
-        bar.setValue(bar.value() + bar.singleStep() * units)
+        bar.setValue(int(bar.value() + bar.singleStep() * units))
 
     def setZoom(self, value):
         self.actions.fitWidth.setChecked(False)
